Patch from Lauri H�stholm:
--
Another thing I recently noticed is that fcpDestroyHandle() doesn't seem
to close sockets. Because of this you have to use both fcpCloseKey() and
the destroy call when you're done with a handle. However, when key opening
fails and you call fcpCloseKey(), you may eventually end up with the open
socket counter showing negative values. I have attached a patch that adds
a small check to _fcpSockDisconnect() in _fcpSock.c and to
fcpDestroyHandle.c. The patch makes the counter only be decreased if a
socket is actually closed and on the other hand it's not necessary to
close keys anymore as long as you destroy the handle. This might even
prevent memory leaks in some cases.
--
Lauri


git-svn-id: file:///home/toad/git-migration/temprepository/trunk/contrib@1084 67a373e5-eb02-0410-a15c-ee090a768436